remove warning

+++ src/gdxviewer/filteruelmodel.cpp
@@ -58,7 +58,7 @@ int FilterUelModel::rowCount(const QModelIndex &parent) const
     // other (valid) parents, rowCount() should return 0 so that it does not become a tree model.
     if (parent.isValid())
         return 0;
-    return mUels->size();
+    return (int) mUels->size();
 }
 
 QVariant FilterUelModel::data(const QModelIndex &index, int role) const
